update notification module

+++ Modules/Notification/Database/Migrations/2021_11_13_000000_create_notification_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateSmsLogTable extends Migration
+class CreateNotificationTable extends Migration
 {
     /**
      * Run the migrations.
@@ -23,6 +23,24 @@ class CreateSmsLogTable extends Migration
                 $table->tinyInteger('status')->default(1);
                 $table->timestamps();
                 $table->softDeletes();
+                $table->integer('created_by')->default(0);
+                $table->foreign('com_id')->references('id')->on('companies')->onDelete('SET NULL');
+                $table->foreign('branch_id')->references('id')->on('branches')->onDelete('SET NULL');
+                $table->foreign('employee_id')->references('id')->on('employees')->onDelete('SET NULL');
+            });
+        }
+
+        if (! Schema::hasTable('email_log')) {
+            Schema::create('email_log', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger('com_id')->unsigned()->nullable();
+                $table->bigInteger('branch_id')->unsigned()->nullable();
+                $table->bigInteger('employee_id')->unsigned()->nullable();
+                $table->string('subject')->nullable();
+                $table->string('email')->nullable();
+                $table->json('body')->nullable();
+                $table->tinyInteger('status')->default(1);
+                $table->timestamps();
                 $table->integer('created_by')->default(0);
                 $table->foreign('com_id')->references('id')->on('companies')->onDelete('SET NULL');
                 $table->foreign('branch_id')->references('id')->on('branches')->onDelete('SET NULL');
